refactor custom property format
